Split up test_class into multiple files
